Added more detail to README and added bash directory
Main README now has some more detail; I added some links to
some additional helpful tools (i.e. Anaconda and Postman), and
explained the precise scope of this repo's coverage.

In particular, explained that it only covers things that either
allow for customization that I've already set for my environment,
or have a number of issues/complicated instructions that would be helpful
to record for the future. Will only link to/provide instructions for
simpler things if they are prerequisites for the actual things listed.

Thus, it won't cover something like installing Python, since that's easy
to install and usually pretty necessary for a dev environment in the first
place.

Also added directory for bash; currently just links to several bash
prompts on GitHub that are nice (like Sexy Bash Prompt or
Powerline-variants).
	new file:   Bash/README.md
	modified:   README.md